pyproject.toml
[project]
name = "blend360"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "duckdb",
    "faiss-cpu>=1.13.0",
    "fastapi>=0.122.0",
    "google-genai>=1.52.0",
    "google-generativeai>=0.8.5",
    "gradio>=6.0.1",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langgraph>=1.0.4",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "sentence-transformers>=5.1.2",
    "sqlparse>=0.5.4",
    "tabulate>=0.9.0",
    "uvicorn[standard]>=0.38.0",
]
